Moved moveInputCallback to the state class. Removed `enableSquareSelect`. Removed state.inputEnabled.
This commit changes the implementation of `inputEnabled` to a method that checks either `inputWhiteEnabled` or `inputBlackEnabled` is enabled. Predicted to provide more accurate responses to certain game states.

Additionally, the `enableSquareSelect` and `disableSquareSelect` methods are deprecated and removed because you can directly add events to `chessboard.context` and read the square from `event.target.getAttribute("data-square")`. The associated code and documentation are removed to improve clarity and reduce maintenance.